Fixed the way database healthcheck works. Improved efficiency of docker-init.sh

+++ sample-configs/docker-init.sh
@@ -3,21 +3,18 @@
 if [ -d "../../database" ]; then
 	echo "Database directory exists"
 else
-	git clone git@bitbucket.org:geneweaver/database.git ../../database
-	cd ../../database
-	git fetch && git checkout docker
-	cd ../website-py/sample-configs
+	git clone -b docker git@bitbucket.org:geneweaver/database.git ../../database
 fi
 
 
 if [ -d "../../tools" ]; then
 	echo "Tools directory exists"
 else
-	git clone git@bitbucket.org:geneweaver/tools.git ../../tools
-	cd ../../tools
-	git fetch && git checkout docker
-	cd ../website-py/sample-configs
+	git clone -b docker git@bitbucket.org:geneweaver/tools.git ../../tools
 fi
 
-docker volume rm $(docker volume ls -qf dangling=true)
+if [ $(docker volume ls -qf dangling=true | wc -w) -gt 0 ]; then
+	docker volume rm $(docker volume ls -qf dangling=true)
+fi
+
 docker-compose build
